+++ noxfile.py
@@ -62,10 +62,9 @@ def integration(session):
 
 @nox.session
 def examples(session):
-    """Run the examples and notebooks"""
+    """Run the example scripts."""
     session.install("-e", ".[all,dev]", "--upgrade", silent=False)
     session.run("pytest", "--examples")
-    notebooks(session)
 
 
 @nox.session
@@ -150,6 +149,7 @@ def run_quick(session):
     """
     run_tests(session)
     run_doc_tests(session)
+    notebooks(session)
 
 
 @nox.session