+++ pybop/__init__.py
@@ -48,7 +48,7 @@ from ._utils import add_spaces, is_numeric, FailedVariable, FailedSolution, Symb
 #
 # Dataset class
 #
-from ._dataset import Dataset
+from ._dataset import Dataset, import_pyprobe_result
 
 #
 # Transformation classes