+++ tests/unit/test_plots.py
@@ -105,7 +105,9 @@ class TestPlots:
         pybop.plot.problem(design_problem)
 
         # Test conversion of values into inputs
-        pybop.plot.problem(fitting_problem, problem_inputs=[0.6, 0.6])
+        pybop.plot.problem(
+            fitting_problem, inputs=fitting_problem.parameters.to_dict([0.6, 0.6])
+        )
 
     def test_cost_plots(self, fitting_problem):
         # Test plot of Cost objects
@@ -170,8 +172,8 @@ class TestPlots:
         problem = pybop.Problem(simulator, posterior)
         options = pybop.PintsSamplerOptions(n_chains=1, max_iterations=1)
         sampler = pybop.SliceStepoutMCMC(problem, options=options)
-        results = sampler.run()
-        return pybop.PosteriorSummary(results)
+        result = sampler.run()
+        return pybop.PosteriorSummary(result.chains)
 
     def test_posterior_plots(self, posterior_summary):
         # Plot trace
@@ -286,7 +288,7 @@ class TestPlots:
 
         # Plot the nyquist
         inputs = problem.parameters.to_dict([60e-6])
-        pybop.plot.nyquist(problem, problem_inputs=inputs, title="Optimised Comparison")
+        pybop.plot.nyquist(problem, inputs=inputs, title="Optimised Comparison")
 
         # Without inputs
         pybop.plot.nyquist(problem, title="Optimised Comparison")